Refactor add custom field form to separate component
[MAILPOET-2463]
